# `<regex>`: Improve search performance for regexes with initial `+` quantifiers